Updated leads endpoints to take $options array
Instead of taking separate params (i.e. for the update leads program status endpoint, having people pass in an array of lead ids and another param for the status the lead should be updated to), changed the leads methods to take an array of options. Each option is appended to the body of the request.