Allows for more flexibility with endpoint options
Prior to this, I was explicitly defining the options that could be passed in. Instead of that, I'm now just iterating over the options array passed in and adding them to the URL, regardless of what the params are, so the package user can use whatever options Marketo's API is currently allowing.